Update view.py
11/05/2025

+++ FBLA 2025/Website/view.py
@@ -1,69 +1,79 @@
 from flask import Blueprint, render_template, request, flash, jsonify
-from flask_login import login_user, login_required, logout_user,current_user
-from .models import Note , Task, Journal
+from flask_login import login_user, login_required, logout_user, current_user
+from .models import Note, Task, User, Business
 from Website import db
 import json
 
-views =Blueprint('views', __name__, template_folder= 'Templates')
+views = Blueprint('views', __name__, template_folder='Templates')
 
 
 @views.route('/about')
 def about():
-    return render_template('about.html', user = current_user)
+    return render_template('about.html', user=current_user)
 
-@views.route('/profile')
+
+@views.route('/profile', methods=['GET', 'POST'])
+@login_required
 def profile():
-    return render_template('profile.html', user = current_user)
+    if request.method == 'POST':
+        biz_name = request.form.get('biz_name')
+        biz_desc = request.form.get('title')
+
+        # If you plan to save a business entry, uncomment and adjust below:
+        # if biz_name and biz_desc:
+        #     new_business = Business(name=biz_name, description=biz_desc, user_id=current_user.id)
+        #     db.session.add(new_business)
+        #     db.session.commit()
+        #     flash('Business info saved!', category='success')
+
+    return render_template('profile.html', user=current_user)
+
 
 @views.route('/', methods=['GET', 'POST'])
 @login_required
 def home():
-    if request.method == 'POST': 
-        note = request.form.get('note')#Gets the note from the HTML 
+    if request.method == 'POST':
+        note = request.form.get('note')
         title = request.form.get('title')
-        
+
         if len(note) < 1:
-            flash('Note is too short!', category='error') 
+            flash('Note is too short!', category='error')
         else:
-            new_note = Note(data=note,title=title ,user_id=current_user.id,)  #providing the schema for the note 
-            db.session.add(new_note) #adding the note to the database 
+            new_note = Note(data=note, title=title, user_id=current_user.id)
+            db.session.add(new_note)
             db.session.commit()
             flash('Note added!', category='success')
 
     return render_template("home.html", user=current_user)
 
 
-
-
-
-@views.route('/tasks',methods=['GET', 'POST'])
+@views.route('/tasks', methods=['GET', 'POST'])
 @login_required
 def tasks():
-    if request.method == 'POST': 
-        task = request.form.get('task')#Gets the note from the HTML 
+    if request.method == 'POST':
+        task = request.form.get('task')
         title = request.form.get('title')
-        
+
         if len(task) < 1:
-            flash('Note is too short!', category='error') 
+            flash('Task is too short!', category='error')
         else:
-            new_task = Task(data = task ,title=title ,user_id=current_user.id)  #providing the schema for the note 
-            db.session.add(new_task) #adding the task to the database 
+            new_task = Task(data=task, title=title, user_id=current_user.id)
+            db.session.add(new_task)
             db.session.commit()
             flash('Task added!', category='success')
 
-    return render_template('tasks.html', user = current_user)
+    return render_template('tasks.html', user=current_user)
 
 
 @views.route('/delete-note', methods=['POST'])
-def delete_note():  
-    note = json.loads(request.data) # this function expects a JSON from the INDEX.js file 
+@login_required
+def delete_note():
+    note = json.loads(request.data)  # expects JSON from frontend
     noteId = note['noteId']
     note = Note.query.get(noteId)
-    if note:
-        if note.user_id == current_user.id:
-            flash('Note deleted', category='error')
-            print('note deleted')
-            db.session.delete(note)
-            db.session.commit()
-            
+    if note and note.user_id == current_user.id:
+        db.session.delete(note)
+        db.session.commit()
+        flash('Note deleted', category='success')
+
     return jsonify({})
